+++ yash-env/src/semantics.rs
@@ -18,6 +18,7 @@
 
 use crate::signal;
 use crate::system::Errno;
+use crate::system::System;
 use nix::sys::signal::Signal;
 use std::ffi::c_int;
 use std::ops::ControlFlow;
@@ -104,9 +105,29 @@ impl From<ExitStatus> for c_int {
 ///
 /// POSIX requires the exit status to be greater than 128. The current
 /// implementation returns `signal_number + 384`.
+///
+/// See [`ExitStatus::to_signal_number`] for the reverse conversion.
 impl From<signal::Number> for ExitStatus {
     fn from(number: signal::Number) -> Self {
         Self::from(number.as_raw() + 0x180)
+    }
+}
+
+impl ExitStatus {
+    /// Returns the signal number corresponding to the exit status.
+    ///
+    /// Basically, this function is the inverse of the `From<signal::Number>`
+    /// implementation for `ExitStatus`. It tries to find a signal number by
+    /// offsetting the exit status by 384. If the offsetting does not result in a
+    /// valid signal number, it additionally tries with 128 and 0.
+    #[must_use]
+    pub fn to_signal_number<S: System>(self, system: &S) -> Option<signal::Number> {
+        [0x180, 0x80, 0]
+            .into_iter()
+            .filter_map(|offset| self.0.checked_sub(offset))
+            .filter_map(|raw_number| system.validate_signal(raw_number))
+            .next()
+            .map(|(_name, number)| number)
     }
 }
 
@@ -243,6 +264,8 @@ pub type Result<T = ()> = ControlFlow<Divert, T>;
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::system::r#virtual::VirtualSystem;
+    use crate::system::r#virtual::{SIGINT, SIGTERM};
 
     #[test]
     fn signal_try_from_exit_status() {
@@ -262,5 +285,28 @@ mod tests {
             Signal::try_from(ExitStatus::from(Signal::SIGHUP)),
             Ok(Signal::SIGHUP)
         );
+    }
+
+    #[test]
+    fn exit_status_to_signal_number() {
+        let system = VirtualSystem::new();
+
+        assert_eq!(ExitStatus(0).to_signal_number(&system), None);
+
+        assert_eq!(
+            ExitStatus(SIGINT.as_raw()).to_signal_number(&system),
+            Some(SIGINT)
+        );
+        assert_eq!(
+            ExitStatus::from(SIGINT).to_signal_number(&system),
+            Some(SIGINT)
+        );
+
+        let mut exit_status = ExitStatus::from(SIGTERM);
+        assert_eq!(exit_status.to_signal_number(&system), Some(SIGTERM));
+        exit_status.0 &= 0xFF;
+        assert_eq!(exit_status.to_signal_number(&system), Some(SIGTERM));
+        exit_status.0 &= 0x7F;
+        assert_eq!(exit_status.to_signal_number(&system), Some(SIGTERM));
     }
 }